Bugfixes for SHIVs and psi soldiers
- Don't apply a loadout to SHIVs in the squad loadout
- Don't touch custom items when applying loadouts.

The custom item one may introduce other bugs, but it looks like these are set generally at the
time a soldier has their armor equipped and when they are loaded into the skyranger - grapple
hooks and special psi amp devices, respectively. So don't take any special action with this and
let the normal skyranger loading code handle the psi-amp (which is loadout independent) and the
armor equip code handle the grapples, which should apply regularly by the loadout in the mod.