Disable loadout save/restore menus on SHIVs.
SHIVs were already excluded from squad restore with an earlier commit,
but it was still possible to use the save/restore menu when viewing a
SHIV in the individual soldier screen. SHIVs were incorrectly considered
to be PFCs and restoring a PFC loadout onto a SHIV would corrupt the
unit. Similarly saving a SHIV loadout and applying it to a PFC could do
bad things.

For now, just disable the menu on SHIVs. They could be supported by
adding their own distinct banks, but this may require versioning the
checkpoint data so as to remain compatible with existing saves. I'm
not sure if the checkpoint supports loading from an existing save with a
slightly different structure.